+++ pkg/management/postgres/join.go
@@ -33,16 +33,9 @@ import (
 // ClonePgData clones an existing server, given its connection string,
 // to a certain data directory
 func ClonePgData(connectionString, targetPgData, walDir string) error {
-	// To initiate streaming replication, the frontend sends the replication parameter
-	// in the startup message. A Boolean value of true (or on, yes, 1) tells the backend
-	// to go into physical replication walsender mode, wherein a small set of replication
-	// commands, shown below, can be issued instead of SQL statements.
-	// https://www.postgresql.org/docs/current/protocol-replication.html
-	connectionString += " replication=1"
-
 	log.Info("Waiting for server to be available", "connectionString", connectionString)
 
-	db, err := pool.NewDBConnection(connectionString, pool.ConnectionProfilePostgresql)
+	db, err := pool.NewDBConnection(connectionString, pool.ConnectionProfilePostgresqlPhysicalReplication)
 	if err != nil {
 		return err
 	}
