+++ pkg/management/postgres/pool/connection.go
@@ -30,45 +30,6 @@ type ConnectionProfile interface {
 	Enrich(config *pgx.ConnConfig)
 }
 
-var (
-	// ConnectionProfilePostgresql is the connection profile to be used for PostgreSQL
-	ConnectionProfilePostgresql ConnectionProfile = connectionProfilePostgresql{}
-
-	// ConnectionProfilePgbouncer is the connection profile to be used for Pgbouncer
-	ConnectionProfilePgbouncer ConnectionProfile = connectionProfilePgbouncer{}
-)
-
-type connectionProfilePostgresql struct{}
-
-func (connectionProfilePostgresql) Enrich(config *pgx.ConnConfig) {
-	// This is required by pgx when using the simple protocol during
-	// the sanitization of the strings. Do not remove.
-	config.RuntimeParams["client_encoding"] = "UTF8"
-
-	// Set the default datestyle in the connection helps to keep
-	// a standard date format for the operator to manage the dates
-	// when it's needed
-	config.RuntimeParams["datestyle"] = "ISO"
-}
-
-type connectionProfilePgbouncer struct{}
-
-func (connectionProfilePgbouncer) Enrich(config *pgx.ConnConfig) {
-	// The simple query protocol is needed since we're going to use
-	// this function to connect to the PgBouncer administrative
-	// interface, which doesn't support the extended one.
-	config.DefaultQueryExecMode = pgx.QueryExecModeSimpleProtocol
-
-	// This is required by pgx when using the simple protocol during
-	// the sanitization of the strings. Do not remove.
-	config.RuntimeParams["client_encoding"] = "UTF8"
-
-	// Set the default datestyle in the connection helps to keep
-	// a standard date format for the operator to manage the dates
-	// when it's needed
-	config.RuntimeParams["datestyle"] = "ISO"
-}
-
 // NewDBConnection creates a postgres connection with the simple protocol
 func NewDBConnection(connectionString string, profile ConnectionProfile) (*sql.DB, error) {
 	conf, err := pgx.ParseConfig(connectionString)
